+++ R/chrome.R
@@ -95,10 +95,13 @@ chrome_print = function(
   ))
 
   debug_port = servr::random_port()
+  log_file = if (isTRUE(getOption('pagedown.keep.chrome.log'))) {
+    sprintf('chrome-stderr-%i.log', as.integer(Sys.time()))
+  }
   ps = processx::process$new(browser, c(
     paste0('--remote-debugging-port=', debug_port),
     paste0('--user-data-dir=', work_dir), extra_args
-  ))
+  ), stderr = log_file)
   kill_chrome = function(...) {
     if (verbose >= 1) message('Closing browser')
     if (ps$is_alive()) ps$kill()