+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Document that the `html_letter()` output format modifies the default value of the `fig_caption` parameter and allow the user to change it (thanks, @NewGraphEnvironment, #234).
 
+- Created an option `pagedown.keep.chrome.log` to ease `chrome_print()` debugging: when set to `TRUE`, the logs of Chrome are saved in the working directory.
+
 # CHANGES IN pagedown VERSION 0.15
 
 ## NEW FEATURES