+++ NEWS.md
@@ -6,6 +6,8 @@
 
 - Created an option `pagedown.chrome.log` to ease `chrome_print()` debugging: when set to `TRUE`, the logs of Chrome are saved in the working directory.
 
+- In `chrome_print()`, set a timeout when R tries to detect headless Chrome.
+
 # CHANGES IN pagedown VERSION 0.15
 
 ## NEW FEATURES