+++ NEWS.md
@@ -8,7 +8,7 @@
 
 - In `chrome_print()`, set a timeout when R tries to detect headless Chrome.
 
-- Added support for Sass files.
+- Added support for Sass files with `.scss` and `.sass` extension (thanks, @felipecrp, #244).
 
 # CHANGES IN pagedown VERSION 0.15